Rename to spi_led_strip
