+++ esphome/components/spi_led_strip/__init__.py
