+++ tests/test8.yaml
@@ -29,7 +29,7 @@ light:
     name: neopixel-enable
     internal: false
     restore_mode: ALWAYS_OFF
-  - platform: spi_rgb_led
+  - platform: spi_led_strip
     num_leds: 4
     color_correct: [80%, 60%, 100%]
     id: rgb_led
